+++ src/mcga/brainfuck/Trace.java
@@ -24,7 +24,7 @@ public class Trace extends Interpreter {
     @Override
     public void interpretation(int i){
         super.interpretation(i);
-            logFileStream.println("Exec step => " + Integer.toString(Parser.EXEC_POS)
+            logFileStream.println("Exec step => " + Long.toString(Parser.EXEC_POS)
                     + '\t' + '\t' + "Data pointer loc => " + Brainfuck.memory.getCurrentIndex()
                     + '\t' + '\t' + "Memory => " + Brainfuck.memory);
